feat : 진처도 조회

+++ backend/src/main/java/com/pj/planbee/service/CalendarService.java
@@ -7,7 +7,7 @@ import com.pj.planbee.dto.CalendarDTO;
 
 public interface CalendarService {
 
-   public int getProgress(String calDate, String sessionId);
+   public double getProgress(String calDate, String sessionId);
    
    public Map<String, Integer> curProgress(String userId);
    
@@ -18,4 +18,6 @@ public interface CalendarService {
    
    public void checkMonthly(int year, int month, String userId);
    List<CalendarDTO> getMonthly(int year, int month, String userId, String fileterId);
+   
+   
 }
